third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/script-tag-with-source-implied-host-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?clutter=/xssAuditor/resources/xss.js%3e%3c/script%3e&q=%3cscript%20src=https:127.0.0.1:8443/security' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
